Modularize analyze_alluvial.py into compute + 2 plot scripts (#46)
Split the 1400-line monolithic script into three focused scripts:
- compute_alluvial.py: Steps 0-5 + robustness + lexical validation
  Writes: tab_alluvial.csv, tab_breakpoints.csv, tab_breakpoint_robustness.csv,
          cluster_labels.json, tab_core_shares.csv (new), tab_lexical_tfidf.csv
- plot_fig_breakpoints.py: reads CSVs, writes fig_breakpoints.png
- plot_fig_alluvial.py: reads CSVs, writes fig_alluvial.png + .html

Light-touch refactor: SVG helpers moved to top of plot_fig_alluvial.py,
function definitions kept where they logically belong, COP_EVENTS duplicated
in the two scripts that need it (no new shared module).

New output: tab_core_shares.csv (core paper counts per period/cluster),
needed by plot_fig_alluvial.py to render "% core" annotations without
reloading refined_works.csv.

analyze_alluvial.py replaced with a thin wrapper calling all three scripts
in sequence, so existing usage continues to work.

Makefile: replace single &: grouped rule with separate rules per output,
fixing the dependency problem where fig_composition.png (manuscript) required
running the full alluvial pipeline instead of just compute_alluvial.py.
